Ensure the map will always have a white background
The topography layer has its alpha set to 0.1 to subdue the colors (far too bold to leave it as is).  But we don’t actually need the transparency for the map like we do for the NEXRAD frames, so disabling transparency when saving the map ensures the map doesn’t let the color of its container show through.