Enhancing example HTML
Map is now set in the HTML file with the frames, instead of in the CSS.

Also, image preload tags were added and image tags removed, but the JavaScript will re-generate the image tags programmatically.  This is purely a convenience for users tinkering with the example so they don’t have to edit the frame image names twice.

Lastly, extensive commenting was added to the JavaScript to help users more quickly understand how it works so that it can be readily adapted to their own needs.