Don’t highlight counties in CWA
Determing the County Warning Area for a given Weather Forecast Office is kind of a pain, and it has to be done using the WFO ID, not the ICAO (referred to in code as the `site_id`).

For a lot of radar sites, the WFO is just the last three characters of the ICAO, but not all.  For instance, in San Angelo, the ICAO is “KSJT” and the WFO ID is “SJT”, but for Lubbock, the ICAO is “KLBB” while the WFO ID is “LUB.”  So simply grabbing the last three characters, as I was doing, was wrong.

I tried to figure out a way to determine the WFO ID from a given ICAO (in `RadarLoopGenerator._check_site_coords()`), but I gave up.  I’m sure there’s a way, I’m just not willing to put more effort into it (not right now, anyway).

For now, the easiest thing is to just not care about highlighting CWA counties — we get to knock off a large chunk of code, and we don’t have to worry about finding the WFO ID anymore.