This does a _bunch_
It modernizes the build to not use setup.py, as apparently that's
deprecated. It changes the imports so that they'll work relatively.
It sets up the installation of a 'mr_radar' script so that you can just
call `mr_radar` from the command line.

It also builds a docker container using the Dockerfile that packages all
this up.